Change from int to size_t

+++ src/kernels.cu
@@ -57,7 +57,7 @@ __global__ void init_constants(double *__restrict__ a,
     double *__restrict__ d,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -77,7 +77,7 @@ __global__ void init_randoms(double *__restrict__ a,
     unsigned long long seed)
 {
 
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -96,7 +96,7 @@ __global__ void init_randoms(double *__restrict__ a,
 
 __global__ void initCuda(double *__restrict__ b, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -107,7 +107,7 @@ __global__ void initCuda(double *__restrict__ b, int scalar, const size_t N)
 
 __global__ void initCuda_vec2(double *__restrict__ b, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 2) {
     return;
@@ -125,7 +125,7 @@ __global__ void initCuda_vec2(double *__restrict__ b, int scalar, const size_t N
 
 __global__ void initCuda_vec4(double *__restrict__ b, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 4) {
     return;
@@ -149,7 +149,7 @@ __global__ void initCuda_vec4(double *__restrict__ b, int scalar, const size_t N
 
 __global__ void copyCuda(double *__restrict__ c, double *__restrict__ a, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -161,7 +161,7 @@ __global__ void copyCuda(double *__restrict__ c, double *__restrict__ a, const s
 __global__ void copyCuda_vec2(
     double *__restrict__ c, double *__restrict__ a, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 2) {
     return;
@@ -177,7 +177,7 @@ __global__ void copyCuda_vec2(
 __global__ void copyCuda_vec4(
     double *__restrict__ c, double *__restrict__ a, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 4) {
     return;
@@ -196,7 +196,7 @@ __global__ void copyCuda_vec4(
 
 __global__ void updateCuda(double *__restrict__ a, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -207,7 +207,7 @@ __global__ void updateCuda(double *__restrict__ a, int scalar, const size_t N)
 
 __global__ void updateCuda_vec2(double *__restrict__ a, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 2) {
     return;
@@ -225,7 +225,7 @@ __global__ void updateCuda_vec2(double *__restrict__ a, int scalar, const size_t
 
 __global__ void updateCuda_vec4(double *__restrict__ a, int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 4) {
     return;
@@ -253,7 +253,7 @@ __global__ void triadCuda(double *__restrict__ a,
     const int scalar,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -268,7 +268,7 @@ __global__ void triadCuda_vec2(double *__restrict__ a,
     const int scalar,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 2) {
     return;
@@ -292,7 +292,7 @@ __global__ void triadCuda_vec4(double *__restrict__ a,
     const int scalar,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 4) {
     return;
@@ -319,7 +319,7 @@ __global__ void triadCuda_vec4(double *__restrict__ a,
 __global__ void daxpyCuda(
     double *__restrict__ a, double *__restrict__ b, const int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -331,7 +331,7 @@ __global__ void daxpyCuda(
 __global__ void daxpyCuda_vec2(
     double *__restrict__ a, double *__restrict__ b, const int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 2) {
     return;
@@ -351,7 +351,7 @@ __global__ void daxpyCuda_vec2(
 __global__ void daxpyCuda_vec4(
     double *__restrict__ a, double *__restrict__ b, const int scalar, const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N / 4) {
     return;
@@ -380,7 +380,7 @@ __global__ void striadCuda(double *__restrict__ a,
     double *__restrict__ d,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -395,7 +395,7 @@ __global__ void striadCuda_vec2(double *__restrict__ a,
     double *__restrict__ d,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -420,7 +420,7 @@ __global__ void striadCuda_vec4(double *__restrict__ a,
     double *__restrict__ d,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -450,7 +450,7 @@ __global__ void sdaxpyCuda(double *__restrict__ a,
     double *__restrict__ c,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -464,7 +464,7 @@ __global__ void sdaxpyCuda_vec2(double *__restrict__ a,
     double *__restrict__ c,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -487,7 +487,7 @@ __global__ void sdaxpyCuda_vec4(double *__restrict__ a,
     double *__restrict__ c,
     const size_t N)
 {
-  int tidx = threadIdx.x + blockIdx.x * blockDim.x;
+  size_t tidx = threadIdx.x + blockIdx.x * blockDim.x;
 
   if (tidx >= N) {
     return;
@@ -511,7 +511,7 @@ __global__ void sdaxpyCuda_vec4(double *__restrict__ a,
   }
 }
 
-__device__ void warpReduce(volatile int *shared_data, int tidx)
+__device__ void warpReduce(volatile int *shared_data, size_t tidx)
 {
   shared_data[tidx] += shared_data[tidx + 32];
   shared_data[tidx] += shared_data[tidx + 16];
@@ -528,8 +528,8 @@ __global__ void sumCuda(
 {
   extern __shared__ int shared_data[];
 
-  unsigned int tidx = threadIdx.x;
-  unsigned int i    = blockIdx.x * (blockDim.x * 2) + threadIdx.x;
+  size_t tidx = threadIdx.x;
+  size_t i    = blockIdx.x * (blockDim.x * 2) + threadIdx.x;
   shared_data[tidx] = a[i] + a[i + blockDim.x];
   __syncthreads();
